// lib/ai/types.ts
export type AstroLanguage =
  | "english"
  | "hindi"
  | "marathi"
  | "tamil"
  | "telugu"
  | "nepali";

export type AIProvider =
  | "openai";


export type AIModel =
  | "gpt-5.5"
  | "gpt-5"
  | "gpt-5-mini";


export type AIContentType =
  | "news"
  | "breaking-news"
  | "editorial"
  | "horoscope"
  | "seo"
  | "social"
  | "rewrite"
  | "translate"
  | "summary";


//////////////////////////////////////////////////////////////
// BASE REQUEST
//////////////////////////////////////////////////////////////

export interface AIRequest {

  provider?: AIProvider;

  model?: AIModel;

  language?: AstroLanguage;

  temperature?: number;

  topP?: number;

  maxOutputTokens?: number;

}
// ...
export interface NewsRequest extends AIRequest {

  title: string;

  category: string;

  keywords?: string[];

  location?: string;

  source?: string;

}


//////////////////////////////////////////////////////////////
// EDITORIAL REQUEST
//////////////////////////////////////////////////////////////

export interface EditorialRequest extends AIRequest {

  topic: string;

}


//////////////////////////////////////////////////////////////
// REWRITE REQUEST
//////////////////////////////////////////////////////////////

export interface RewriteRequest extends AIRequest {

  content: string;

}


//////////////////////////////////////////////////////////////
// SEO REQUEST
//////////////////////////////////////////////////////////////

export interface SEORequest extends AIRequest {

  title: string;

  content: string;

}


//////////////////////////////////////////////////////////////
// SOCIAL REQUEST
//////////////////////////////////////////////////////////////

export interface SocialRequest extends AIRequest {

  title: string;

  content: string;

}


//////////////////////////////////////////////////////////////
// HOROSCOPE REQUEST
//////////////////////////////////////////////////////////////

export interface HoroscopeRequest extends AIRequest {

  zodiacSign: string;

  horoscopeDate: string | Date;

}